requirements.txt
# Core dependencies
azure-identity>=1.15.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.32.0
fastapi>=0.109.2
uvicorn>=0.27.1
httpx>=0.26.0
pydantic>=2.5.3

# Azure OpenAI
# azure-ai-projects==1.0.0b6

# Utilities
requests>=2.31.0
python-multipart>=0.0.7
jinja2>=3.1.3

# Chainlit
chainlit>=0.7.700
asyncio>=3.4.3